docs: sd-input docs [skip chromatic] (#1322)
